moved: dto 디렉터리 이동 (Issue/PauseCouponsResultDTO)
