+++ src/main/java/io/saim/dash/coupon/manage/controller/ManageController.java
@@ -14,7 +14,7 @@ import org.springframework.web.bind.annotation.RestController;
 import io.saim.dash.coupon.common.dto.Coupon.CouponBriefDTO;
 import io.saim.dash.coupon.common.dto.Issue.CouponIssueLogDTO;
 import io.saim.dash.coupon.common.dto.Coupon.RegisteredCouponDTO;
-import io.saim.dash.coupon.common.dto.PauseCouponsResultDTO;
+import io.saim.dash.coupon.common.dto.Issue.PauseCouponsResultDTO;
 import io.saim.dash.coupon.common.model.DUMMY_ServiceUser;
 import io.saim.dash.coupon.manage.dto.IssuedRequestResponseDTO;
 import io.saim.dash.coupon.manage.dto.UpdateUsableStatusRequestDTO;
@@ -100,7 +100,7 @@ public class ManageController {
 		return specCoupon;
 	}
 
-	@PatchMapping("{issue_id}/status")
+	@PatchMapping("/{issue_id}/status")
 	public PauseCouponsResponseDTO updateCouponUsableStatus(
 		@AuthenticationPrincipal DUMMY_ServiceUser user,
 		@PathVariable Long issue_id,
